pyproject.toml
[project]
name = "cryptolight"
version = "0.1.0"
description = "코인 자동매매봇 - signalight에서 진화한 암호화폐 트레이딩 시스템"
requires-python = ">=3.11"
dependencies = [
    "httpx>=0.27",
    "pyjwt>=2.8",
    "python-dotenv>=1.0",
    "pydantic-settings>=2.0",
    "pandas>=2.0",
    "apscheduler>=3.10",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
    "ruff>=0.4",
]
telegram = [
    "python-telegram-bot>=21.0",
]
web = [
    "fastapi>=0.110",
    "uvicorn>=0.29",
]

[project.scripts]
cryptolight = "cryptolight.main:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/cryptolight"]

[tool.ruff]
target-version = "py311"
line-length = 100

[tool.pytest.ini_options]
testpaths = ["tests"]
